Move get_device_code outside of Seedr client
