+++ seedrcc/__init__.py
@@ -1,11 +1,13 @@
 from . import exceptions, models
+from .auth import get_device_code
 from .client import Seedr
 from .token import Token
 
 __all__ = [
     "Seedr",
+    "Token",
+    "get_device_code",
     "models",
     "exceptions",
-    "Token",
 ]
