+++ seedrcc/client.py
@@ -43,20 +43,6 @@ class Seedr:
         """The current authentication token used by the client."""
         return self._token
 
-    @staticmethod
-    def get_device_code() -> models.DeviceCode:
-        """
-        Step 1 of the device flow.
-        Gets the device and user codes required for authorization.
-
-        Example:
-            >>> codes = Seedr.get_device_code()
-            >>> print(f"Go to {codes.verification_url} and enter {codes.user_code}")
-        """
-        with httpx.Client() as client:
-            response_data = _auth.get_device_code(client)
-        return models.DeviceCode.from_dict(response_data)
-
     @classmethod
     def from_device_code(
         cls: Type["Seedr"],
